+++ apps/render-composer-examples/src/App.tsx
@@ -1,6 +1,11 @@
 import { Animate } from "@hmans/r3f-animate"
-import { Environment, OrbitControls } from "@react-three/drei"
-import { RenderCanvas, RenderPipeline } from "render-composer"
+import { Environment, Loader } from "@react-three/drei"
+import { Suspense, useLayoutEffect } from "react"
+import {
+  RenderCanvas,
+  RenderPipeline,
+  useRenderPipeline
+} from "render-composer"
 import { Object3D } from "three"
 
 const rotate = (o: Object3D, dt: number) => {
@@ -10,34 +15,53 @@ const rotate = (o: Object3D, dt: number) => {
 
 function App() {
   return (
-    <RenderCanvas>
-      <RenderPipeline vignette bloom antiAliasing>
-        <color attach="background" args={["#264653"]} />
-        <Environment preset="sunset" />
+    <>
+      <Loader />
 
-        <directionalLight position={[30, 10, 10]} intensity={1.5} />
+      <RenderCanvas>
+        <RenderPipeline vignette bloom antiAliasing godRays>
+          <Suspense>
+            <color attach="background" args={["#264653"]} />
+            <Environment preset="sunset" />
 
-        <Animate
-          fun={(o, _, { clock }) => {
-            o.position.x = Math.sin(clock.getElapsedTime() * 0.7) * 2
-            o.position.y = Math.sin(clock.getElapsedTime() * 1.1)
-            o.position.z = Math.cos(clock.getElapsedTime() * 0.5)
-          }}
-        >
-          <Animate fun={rotate}>
-            <mesh>
-              <icosahedronGeometry />
-              <meshStandardMaterial
-                color="#E9C46A"
-                metalness={0.5}
-                roughness={0.5}
-              />
-            </mesh>
-          </Animate>
-        </Animate>
-      </RenderPipeline>
-    </RenderCanvas>
+            <directionalLight position={[30, 10, 10]} intensity={1.5} />
+
+            <Sun />
+
+            <Animate
+              fun={(o, _, { clock }) => {
+                o.position.x = Math.sin(clock.getElapsedTime() * 0.7) * 2
+                o.position.y = Math.sin(clock.getElapsedTime() * 1.1)
+                o.position.z = Math.cos(clock.getElapsedTime() * 0.5)
+              }}
+            >
+              <Animate fun={rotate}>
+                <mesh>
+                  <icosahedronGeometry />
+                  <meshStandardMaterial
+                    color="#E9C46A"
+                    metalness={0.5}
+                    roughness={0.5}
+                  />
+                </mesh>
+              </Animate>
+            </Animate>
+          </Suspense>
+        </RenderPipeline>
+      </RenderCanvas>
+    </>
   )
+}
+
+const Sun = () => {
+  const { sun } = useRenderPipeline()
+
+  useLayoutEffect(() => {
+    console.log(sun)
+    sun.position.set(40, 10, -100)
+  }, [])
+
+  return null
 }
 
 export default App