+++ packages/input-composer/src/index.ts
@@ -1,0 +1,5 @@
+export * from "./functions"
+export * from "./interactions"
+export * from "./types"
+export * from "./useGamepad"
+export * from "./useInput"