+++ packages/input-composer/src/types.ts
@@ -1,0 +1,4 @@
+export interface IVector {
+  x: number
+  y: number
+}