+++ packages/shader-composer/src/stdlib/variables.test.ts
@@ -1,5 +1,5 @@
 import { glslRepresentation } from "../glslRepresentation"
-import { UniformUnit } from "./uniforms"
+import { UniformUnit } from "./variables"
 
 describe("Uniform", () => {
   it("uses its uniform name as its string representation", () => {
@@ -13,7 +13,10 @@ describe("Uniform", () => {
   })
 
   it("allows the user to set a uniform name explicitly", () => {
-    const uniform = UniformUnit("float", 0, { variableName: "foo", uniformName: "u_bar" })
+    const uniform = UniformUnit("float", 0, {
+      variableName: "foo",
+      uniformName: "u_bar"
+    })
     expect(glslRepresentation(uniform)).toBe("u_bar")
   })
 })