+++ packages/timeline-composer/src/Repeat.tsx
@@ -5,14 +5,18 @@ type RepeatProps = { children: ReactNode; times?: number; seconds?: number }
 
 const Wrap = ({ children }: { children: ReactNode }) => <>{children}</>
 
-export const Repeat: FC<RepeatProps> = ({ children, times = Infinity, seconds = 1 }) => {
+export const Repeat: FC<RepeatProps> = ({
+  children,
+  times = Infinity,
+  seconds = 1
+}) => {
   const [iteration, setIteration] = useState(1)
 
   return (
-    <Fragment key={iteration}>
+    <Fragment>
       {iteration < times && (
         <Delay
-          key={iteration}
+          key={`delay-${iteration}`}
           seconds={seconds}
           onComplete={() => {
             setIteration((i) => i + 1)