+++ packages/vfx-composer/src/createParticleUnits.ts
@@ -1,4 +1,4 @@
-import { Div, Input, SplitVector2, Sub } from "shader-composer"
+import { Div, Input, Sub, Vec2 } from "shader-composer"
 
 export type ParticleUnits = ReturnType<typeof createParticleUnits>
 
@@ -6,7 +6,8 @@ export const createParticleUnits = (
   lifetime: Input<"vec2">,
   time: Input<"float">
 ) => {
-  const [startTime, endTime] = SplitVector2(lifetime)
+  const { x: startTime, y: endTime } = Vec2(lifetime)
+
   const maxAge = Sub(endTime, startTime)
   const age = Sub(time, startTime)
   const progress = Div(age, maxAge)