+++ packages/material-composer/CHANGELOG.md
@@ -1,5 +1,31 @@
 # material-composer
 
+## 0.2.1
+
+### Patch Changes
+
+- e632a35: The `Rotate` module will now also, by default, rotate the vertex normal. If, for some reason, you don't want this to happen, just set the `normal` option to `false`.
+- e632a35: The `Rotate` module now also happily accepts rotation matrices of type `mat4`.
+
+## 0.2.0
+
+### Minor Changes
+
+- 5af254e: The default for the `space` prop of `Translate` has been changed from `world` to `local`.
+
+### Patch Changes
+
+- Updated dependencies [a92d0d3]
+- Updated dependencies [433f93b]
+- Updated dependencies [c3dcb12]
+- Updated dependencies [765b29d]
+- Updated dependencies [9406986]
+- Updated dependencies [6d99c19]
+- Updated dependencies [765b29d]
+  - shader-composer@0.4.0
+  - @material-composer/patch-material@0.1.2
+  - shader-composer-toybox@0.1.3
+
 ## 0.1.3
 
 ### Patch Changes