+++ packages/vfx-composer/CHANGELOG.md
@@ -1,5 +1,19 @@
 # vfx-composer
 
+## 0.2.3
+
+### Patch Changes
+
+- Updated dependencies [a92d0d3]
+- Updated dependencies [433f93b]
+- Updated dependencies [765b29d]
+- Updated dependencies [9406986]
+- Updated dependencies [5af254e]
+- Updated dependencies [6d99c19]
+- Updated dependencies [765b29d]
+  - shader-composer@0.4.0
+  - material-composer@0.2.0
+
 ## 0.2.2
 
 ### Patch Changes